+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 # 1.0.6 Hotfix and adjustmets
 # 2.0.1 Renaming and improvements
 # 3.0.0 Renaming Assembler Definitions
+# 4.0.0 Code improvement
+- The STransform class was renamed to TransformHandler.
+- The 'Verify' methods were created to remove code duplication.
+- The 'UpdateValues' method was created so that the proportion values of the rect in rect transform are updated.
+- Handling of return values from RectTransform manipulation was created.
+- Dependencies for modules were created for better visualization and manipulation of attributes in the Inspector.
+- Code improvement.